Move widgets to a separate package
