+++ app/src/main/java/com/kamron/pogoiv/widgets/PokemonSpinnerAdapter.java
@@ -1,4 +1,4 @@
-package com.kamron.pogoiv;
+package com.kamron.pogoiv.widgets;
 
 import android.content.Context;
 import android.view.LayoutInflater;
@@ -7,6 +7,7 @@ import android.view.ViewGroup;
 import android.widget.ArrayAdapter;
 import android.widget.TextView;
 
+import com.kamron.pogoiv.GUIUtil;
 import com.kamron.pogoiv.logic.Pokemon;
 
 import java.util.ArrayList;
